remove init cmd, add registry cmd
