+++ packages/teraslice-cli/src/cmds/assets/index.ts
@@ -2,7 +2,7 @@ import { CMD } from '../../interfaces.js';
 import build from './build.js';
 import deleteCmd from './delete.js';
 import deploy from './deploy.js';
-import init from './init.js';
+import registry from './registry.js';
 import list from './list.js';
 
 export default {
@@ -11,7 +11,7 @@ export default {
     exclude: 'lib',
     builder(yargs) {
         return yargs.strict()
-            .command([build, deleteCmd, deploy, init, list])
+            .command([build, deleteCmd, deploy, registry, list])
             .demandCommand(2);
     },
     handler() {}
